% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_click_to_zoom.R
\name{add_click_to_zoom}
\alias{add_click_to_zoom}
\title{Add closeup image overlay}
\usage{
add_click_to_zoom()
}
\description{
Add HTML that enables the user to click an image to bring up a large overlay of that image without needing to go to another page.
}
