% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_endorsements.R
\name{print_endorsements}
\alias{print_endorsements}
\title{Output endorsements from an Rmarkdown file}
\usage{
print_endorsements()
}
\description{
Processes endorsements in the YAML header of an RMarkdown file and outputs them. Designed for use in distill templates
}
